Fix join username not showing for requests
